Organized the task code across packages and modules.
