--- agents_playground/tasks/tasks_registry.py
@@ -1,8 +1,0 @@
-from typing import Callable, Dict, Final
-
-TASKS_REGISTRY: Final[Dict[str, Callable]] = {
-    # 'agent_pacing': am.agent_pacing,
-    # 'agents_spinning' : am.agents_spinning,
-    # 'agent_random_navigation' : am.agent_random_navigation,
-    # 'generate_agents': ga.generate_agents
-}
